pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nanocortex"
version = "0.1.0"
description = "Unified AI System: a reference architecture for auditable multimodal AI decision-making"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25",
    "python-dotenv>=1.0",
    "Pillow>=10.0",
    "PyMuPDF>=1.23",
    "pytesseract>=0.3.10",
    "numpy>=1.24",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
    "fastapi>=0.109",
    "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.2",
    "mypy>=1.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring external services",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
